feat(sidebar): hide terminated pci projects
ref: TAPC-395
